refactor: remove HomeContainer and extract logical function to NavbarContainer
