+++ src/App.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import './App.css'
-import HomeContainer from './components/HomeContainer'
+import Home from './components/Home'
 
 function App() {
   return (
     <div className="App">
-      <HomeContainer />
+      <Home />
     </div>
   )
 }
